requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
python-multipart>=0.0.6
python-dotenv>=1.0.0
slowapi>=0.1.9
google-generativeai>=0.3.0
aiofiles>=23.0.0
jinja2>=3.1.0
psutil>=5.9.0
aiohttp>=3.8.0
lxml_html_clean>=0.4.0